Added datatable in adminProducts
The new datatable includes pagination, sorting by name and price and
page limits. Tables methods and default values are provided in the new
product-table.service. The new SortableTable derectives adds new event
emitter to the th elements in the table header, when click it will emit
the sort event and trigger the sort method in produc-table service